Make RPLidar to headeronly
